Address the issues raised with the initial pull request.
Specifically, any strings found as the "value" of an enum field
are evaled, making them more general.

The loop doesn't require that they are the first in the field
list, and doesn't stop when one is found (allowing mixing of
calculated values and explicit values).

Duplicate values are already checked in _check_enum_value_consistency